+++ ouds_core/lib/components/button/ouds_button.dart
@@ -236,6 +236,7 @@ class _OudsButtonState extends State<OudsButton> {
     );
     final buttonState = buttonStateDeterminer.determineControlState();
     final borderTokens = OudsTheme.of(context).borderTokens;
+    final buttonTokens = OudsTheme.of(context).componentsTokens(context).button;
 
     try {
       if (widget.appearance == OudsButtonAppearance.negative && OudsTheme.isOnColoredSurfaceOf(context)) {
@@ -246,28 +247,36 @@ class _OudsButtonState extends State<OudsButton> {
       debugPrint("Warning: ${e.toString()}");
     }
 
-    return _isFocused
-        ? FocusContainer(
-      color: OudsTheme.of(context).colorScheme(context).borderFocus,
-      strokeWidth: borderTokens.widthFocus,
-      borderRadius: OudsButtonBorderModifier.getDoubleRadiusFocus(context),
-      alignment: FocusAlignment.center,
-      isFocused: _isFocused,
-      child: _buildLayout(
-        context,
-        buttonState,
+    return ConstrainedBox(
+      constraints: BoxConstraints(
+        minWidth: buttonTokens.sizeMinWidth,
       ),
-    )
-        : _buildLayout(
-      context,
-      buttonState,
+      child: SizedBox(
+        height: buttonTokens.sizeMinHeight,
+        child: _isFocused
+            ? FocusContainer(
+                color: OudsTheme.of(context).colorScheme(context).borderFocus,
+                strokeWidth: borderTokens.widthFocus,
+                borderRadius: OudsButtonBorderModifier.getDoubleRadiusFocus(context),
+                alignment: FocusAlignment.center,
+                isFocused: _isFocused,
+                child: _buildLayout(
+                  context,
+                  buttonState,
+                ),
+              )
+            : _buildLayout(
+                context,
+                buttonState,
+              ),
+      ),
     );
   }
 
   Widget _buildLayout(
-      BuildContext context,
-      OudsButtonControlState buttonState,
-      ) {
+    BuildContext context,
+    OudsButtonControlState buttonState,
+  ) {
     switch (widget.layout) {
       case OudsButtonLayout.iconOnly:
         return _buildButtonIconOnly(context, buttonState);
@@ -304,6 +313,7 @@ class _OudsButtonState extends State<OudsButton> {
                         width: buttonToken.spaceColumnGapIcon,
                       ),
                       Flexible(
+                        fit: FlexFit.loose,
                         child: Text(
                           widget.label ?? "",
                           style: TextStyle(
@@ -350,6 +360,7 @@ class _OudsButtonState extends State<OudsButton> {
                             width: buttonToken.spaceColumnGapIcon,
                           ),
                           Flexible(
+                            fit: FlexFit.loose,
                             child: Text(
                               widget.label ?? "",
                               textAlign: TextAlign.center,
@@ -369,6 +380,8 @@ class _OudsButtonState extends State<OudsButton> {
   }
 
   Widget _buildButtonIconOnly(BuildContext context, OudsButtonControlState buttonState) {
+    final buttonToken = OudsTheme.of(context).componentsTokens(context).button;
+
     switch (buttonState) {
       case OudsButtonControlState.loading:
         final buttonIconOnly = Semantics(
@@ -403,11 +416,14 @@ class _OudsButtonState extends State<OudsButton> {
               label: OudsLocalizations.of(context)?.core_button_icon_only_a11y,
               button: true,
               child: ExcludeSemantics(
-                child: IconButton(
-                  focusNode: _focusNode,
-                  style: OudsButtonStyleModifier.buildButtonStyle(context, appearance: widget.appearance, layout: widget.layout, buttonState: buttonState),
-                  onPressed: widget.onPressed == null ? null : () => _handlePressed(widget.onPressed),
-                  icon: _buildIcon(context, widget.icon!, widget.appearance, widget.layout, buttonState),
+                child: SizedBox(
+                  width: buttonToken.sizeMinWidth,
+                  child: IconButton(
+                    focusNode: _focusNode,
+                    style: OudsButtonStyleModifier.buildButtonStyle(context, appearance: widget.appearance, layout: widget.layout, buttonState: buttonState),
+                    onPressed: widget.onPressed == null ? null : () => _handlePressed(widget.onPressed),
+                    icon: _buildIcon(context, widget.icon!, widget.appearance, widget.layout, buttonState),
+                  ),
                 ),
               ),
             ),
@@ -447,7 +463,7 @@ class _OudsButtonState extends State<OudsButton> {
       default:
         final buttonTextOnly = OutlinedButton(
           focusNode: _focusNode,
-          style: OudsButtonStyleModifier.buildButtonStyle(context, appearance: widget.appearance, layout: widget.layout),
+          style: OudsButtonStyleModifier.buildButtonStyle(context, appearance: widget.appearance, layout: widget.layout, buttonState: buttonState),
           onPressed: widget.onPressed == null ? null : () => _handlePressed(widget.onPressed),
           child: Text(
             widget.label ?? "",
@@ -482,19 +498,19 @@ class _OudsButtonState extends State<OudsButton> {
     if (widget.isFullWidth == true) {
       return SizedBox(
         width: double.infinity,
-        child: widget.isFullWidth == false ? Center(child: child) : child,
+        child: child,
       );
     }
     return child;
   }
 
   Widget _buildIcon(
-      BuildContext context,
-      String assetName,
-      final OudsButtonAppearance appearance,
-      final OudsButtonLayout layout,
-      final OudsButtonControlState buttonState,
-      ) {
+    BuildContext context,
+    String assetName,
+    final OudsButtonAppearance appearance,
+    final OudsButtonLayout layout,
+    final OudsButtonControlState buttonState,
+  ) {
     return SvgPicture.asset(
       excludeFromSemantics: true,
       package: widget.package,